Added "Back to Watch Face" Button/App
Implemented a clock button/app within the launcher that, when clicked, simulates a home button press, bringing the user back to the watch face. This is intended to be similar to how the Apple Watch's launcher functions. Overall, it functions as it should, but I may add a few bug fixes to it.