New Feature:
Implemented the ability to zoom out in launcher for easier access to applications. This is beneficial to users who have an immense amount of applications installed on their watch, which would otherwise take awhile to find using the normal view of the launcher.

TO DO: Implement haptic feedback when zooming out in launcher using the bezel.